favorite btn update

+++ src/components/Badge.js
@@ -1,16 +1,28 @@
 import { dbService } from "fbase";
 import React, { useCallback, useEffect, useState } from "react";
-const Badge = ({ favorite, siteName, userObj, titleId }) => {
+const Badge = ({ favorite, siteName, userObj, titleId, day }) => {
   const axios = require("axios");
   const cheerio = require("cheerio");
   let URL = `/webtoon/list.nhn?titleId=${titleId}`;
   const [updateDay, setUpdataDay] = useState("");
   const [isUpdate, setIsUpdate] = useState(false);
-  const addToFirebase = async (event) => {
+  const [isFavorite, setIsFavorite] = useState(false);
+
+  const checkIsFav = useCallback(async () => {
+    const exist = await (
+      await dbService.doc(`${userObj.uid}/${siteName}/${day}/${titleId}`).get()
+    ).exists;
+    setIsFavorite(exist);
+  }, [day, siteName, userObj, titleId]);
+
+  const badgeOnClick = async (event) => {
     const {
       target: {
         parentElement: { parentElement },
       },
+    } = event;
+    const {
+      target: { title },
     } = event;
     const thumbLink = parentElement.firstChild;
     const thumbImg = thumbLink.firstChild;
@@ -24,39 +36,36 @@ const Badge = ({ favorite, siteName, userObj, titleId }) => {
       titleId: id,
       day: day,
     };
-    await dbService
-      .collection(userObj.uid)
-      .doc(siteName)
-      .collection(targetWebtoonObj.day)
-      .doc(targetWebtoonObj.title)
-      .set(targetWebtoonObj);
-  };
-  const delteToFirebase = async (event) => {
-    const {
-      target: {
-        parentElement: { parentElement },
-      },
-    } = event;
-    const thumbLink = parentElement.firstChild;
-    const thumbImg = thumbLink.firstChild;
-    const index = thumbImg.alt.indexOf("/");
-    const day = thumbImg.alt.substring(index + 1);
-    const targetWebtoonObj = {
-      link: thumbLink.href,
-      img: thumbImg.src,
-      title: thumbImg.title,
-      day: day,
-    };
-    const ok = window.confirm(
-      `Are you sure, you want to delete '${targetWebtoonObj.title}' ?`
-    );
-    if (ok) {
-      await dbService
-        .collection(userObj.uid)
-        .doc(siteName)
-        .collection(targetWebtoonObj.day)
-        .doc(targetWebtoonObj.title)
-        .delete();
+    if (title === "pop") {
+      const ok = window.confirm(
+        `Are you sure, you want to delete '${targetWebtoonObj.title}' from Your Profile ?`
+      );
+      if (ok) {
+        await dbService
+          .doc(
+            `${userObj.uid}/${siteName}/${targetWebtoonObj.day}/${targetWebtoonObj.titleId}`
+          )
+          .delete();
+        if (favorite) {
+          setIsFavorite((prev) => !prev);
+        }
+      }
+    } else if (title === "push") {
+      const ok = window.confirm(
+        `Are you sure, you want to Add '${targetWebtoonObj.title}' to Your Profile ?`
+      );
+      if (ok) {
+        await dbService
+          .doc(
+            `${userObj.uid}/${siteName}/${targetWebtoonObj.day}/${targetWebtoonObj.titleId}`
+          )
+          .set(targetWebtoonObj);
+        if (favorite) {
+          setIsFavorite((prev) => !prev);
+        }
+      }
+    } else {
+      console.log("target Name is undefined");
     }
   };
 
@@ -72,20 +81,38 @@ const Badge = ({ favorite, siteName, userObj, titleId }) => {
   useEffect(() => {
     if (favorite === false) {
       updateInfo();
+    } else {
+      checkIsFav();
     }
-  }, [updateInfo, favorite]);
+  }, [updateInfo, favorite, checkIsFav]);
 
   return (
     <span className="favorite-area">
       {favorite ? (
         <>
-          <i className="fas fa-heart" onClick={addToFirebase}></i>
+          {isFavorite ? (
+            <i
+              className="fas fa-heart active"
+              onClick={badgeOnClick}
+              title="pop"
+            ></i>
+          ) : (
+            <i
+              className="far fa-heart non-active"
+              onClick={badgeOnClick}
+              title="push"
+            ></i>
+          )}
         </>
       ) : (
         <>
           {isUpdate && <span className="update-badge">UP</span>}
           <span className="update-day">{updateDay}</span>
-          <i className="far fa-trash-alt" onClick={delteToFirebase}></i>
+          <i
+            className="far fa-trash-alt delete-btn"
+            onClick={badgeOnClick}
+            title="pop"
+          ></i>
         </>
       )}
     </span>
